+++ ClassicalSharp/2D/Screens/Menu/EditHotkeyScreen.cs
@@ -23,7 +23,7 @@ namespace ClassicalSharp.Gui.Screens {
 		
 		public override void Render(double delta) {
 			base.Render(delta);
-			float cX = game.Width / 2, cY = game.Height / 2;
+			int cX = game.Width / 2, cY = game.Height / 2;
 			gfx.Draw2DQuad(cX - 250, cY - 65, 500, 2, grey);
 			gfx.Draw2DQuad(cX - 250, cY + 45, 500, 2, grey);
 		}
